src/tests/test_chain_bot.py
import unittest
from unittest.mock import MagicMock, patch
# from llm_chain.chain_bot import ChatPreprocess, ChainFactory
# How do I solve this import error?
# I know but can you tell me how to solve it?
# I'm not sure, but I think you can solve it by adding the src directory to the PYTHONPATH environment variable.
# How do I do that?
# You can do that by running the following command in your terminal: export PYTHONPATH=$PYTHONPATH:/path/to/your/src/directory
# Can you run it in Python code?
# Sure, you can run it in Python code by using the os.environ dictionary to modify the PYTHONPATH environment variable.
# Can you show me how to do that?
# Sure, here's an example of how to do that:

import os
os.environ['PYTHONPATH'] = os.environ.get('PYTHONPATH', '') + '/src/'
from llm_chain.chain_bot import ChatPreprocess, ChainFactory

class TestChatPreprocess(unittest.TestCase):

    def setUp(self):
        self.tokenizer = MagicMock()
        self.chat_preprocess = ChatPreprocess(tokenizer=self.tokenizer)

    def test_convert_to_chat_hist(self):
        python_str = "user: Hello\nassistant: Hi\nuser: How are you?\nassistant: I'm good"
        expected_result = {
            "messages": [
                {"role": "user", "content": "Hello"},
                {"role": "assistant", "content": "Hi"},
                {"role": "user", "content": "How are you?"},
                {"role": "assistant", "content": "I'm good"}
            ]
        }
        result = self.chat_preprocess.convert_to_chat_hist(python_str)
        self.assertEqual(result, expected_result)

    def test_truncate_chat_history(self):
        chat_hist = [
            {"role": "user", "content": "Hello"},
            {"role": "assistant", "content": "Hi"},
            {"role": "user", "content": "How are you?"},
            {"role": "assistant", "content": "I'm good"}
        ]
        token_count = 10
        expected_result = chat_hist
        result = self.chat_preprocess.truncate_chat_history(chat_hist, token_count)
        self.assertEqual(result, expected_result)

    # Add more test cases for other methods...

class TestChainFactory(unittest.TestCase):

    def setUp(self):
        self.tokenizer = MagicMock()
        self.chain_factory = ChainFactory(tokenizer=self.tokenizer)

    def test_create_chain_with_history(self):
        chain = self.chain_factory.create_chain_with_history()
        # Assert that the chain is created correctly
        self.assertIsNotNone(chain)
        # Add more assertions for the chain configuration...

    def test_create_is_question_about_petsmart_chain(self):
        chain = self.chain_factory.create_is_question_about_petsmart_chain()
        # Assert that the chain is created correctly
        self.assertIsNotNone(chain)
        # Add more assertions for the chain configuration...

    def test_create_retrieve_document_chain(self):
        chain = self.chain_factory.create_retrieve_document_chain()
        # Assert that the chain is created correctly
        self.assertIsNotNone(chain)
        # Add more assertions for the chain configuration...

    def test_create_generate_query_to_retrieve_context_chain(self):
        chain = self.chain_factory.create_generate_query_to_retrieve_context_chain()
        # Assert that the chain is created correctly
        self.assertIsNotNone(chain)
        # Add more assertions for the chain configuration...

    def test_create_relevant_question_chain(self):
        chain = self.chain_factory.create_relevant_question_chain()
        # Assert that the chain is created correctly
        self.assertIsNotNone(chain)
        # Add more assertions for the chain configuration...

    def test_create_irrelevant_question_chain(self):
        chain = self.chain_factory.create_irrelevant_question_chain()
        # Assert that the chain is created correctly
        self.assertIsNotNone(chain)
        # Add more assertions for the chain configuration...

    def test_create_full_chain(self):
        chain = self.chain_factory.create_full_chain()
        # Assert that the chain is created correctly
        self.assertIsNotNone(chain)
        # Add more assertions for the chain configuration...

if __name__ == '__main__':
    unittest.main()